Should be float.
git-svn-id: https://gforge.sci.utah.edu/svn/seg3d2/branches/rbf_project@2137 305cdda6-5ce1-45b3-a98d-dfc68c8b3305

+++ src/Plugins/rbf/Application/Tools/Actions/ActionRBF.cc
@@ -229,12 +229,11 @@ public:
         {
           for(int i=0; i < dim[0]; i++)
           {
-            double val = value[i][j][k];
+            float val = value[i][j][k];
             nrrd_file.write((char*)&val, sizeof(float));
           }
         }
       }
-      
       nrrd_file.close();
     }
   }
